chore(lint): unite eslint rules
