+++ eslint-plugin/src/core/deprecated-class-method.ts
@@ -22,16 +22,21 @@ export interface ESLintDeprecationStaticMethodCfg {
   deprecatedMethod: string;
   /** Function that returns recommended method */
   getReplacementMethod: (expression: ESTree.CallExpression) => ESLintReplacementMethodCfg | string;
+  /** Callback function to indicate rule skip */
+  skipOn?: () => boolean;
 }
 
 type StaticMethodNode = ESTree.MemberExpression & Rule.NodeParentExtension;
 
 /** Builds deprecation rule from {@link ESLintDeprecationStaticMethodCfg} object */
-export function buildRule(config: ESLintDeprecationStaticMethodCfg): Rule.RuleModule {
+export function buildRule(configs: ESLintDeprecationStaticMethodCfg | ESLintDeprecationStaticMethodCfg[]): Rule.RuleModule {
+  configs = Array.isArray(configs) ? configs : [configs];
   const create = (context: Rule.RuleContext): Rule.RuleListener => {
     return {
       MemberExpression(node: StaticMethodNode): null {
-        if (isDeprecatedMethod(node, config)) handleCallExpression(node, context, config);
+        configs.forEach((config) => {
+          if (isDeprecatedMethod(node, config)) handleCallExpression(node, context, config);
+        });
         return null;
       }
     };
@@ -46,6 +51,7 @@ function isDeprecatedMethod(node: StaticMethodNode, config: ESLintDeprecationSta
 }
 
 function handleCallExpression(node: StaticMethodNode, context: Rule.RuleContext, config: ESLintDeprecationStaticMethodCfg): void {
+  if (config.skipOn && config.skipOn()) return;
   const replCfg = config.getReplacementMethod(node.parent as ESTree.CallExpression);
   const message = typeof replCfg === 'string' ? `${config.className}.${replCfg}` : replCfg.message;
   const replacement = typeof replCfg === 'string' ? replCfg : replCfg.replacement;
