+++ eslint-plugin/src/rules/4/all.rules.ts
@@ -1,21 +1,11 @@
-import deprecatedGenerateUid from './deprecated.generate-uid';
-import deprecatedDeepCompare from './deprecated.deep-compare';
-import deprecatedEventUtils from './deprecated.event-utils';
-import deprecatedTraversingQuery from './deprecated.traversing-query';
-import deprecatedToggleableActionParams from './deprecated.toggleable-action-params';
+import deprecatedAliases from './deprecated.aliases';
 
-import deprecatedBaseDecoratorsPath from './deprecated.base-decorators-path';
+import deprecatedMethods from './deprecated.methods';
 
-import deprecatedMediaRuleListParse from './deprecated.media-rule-list-parse';
+import deprecatedPaths from './deprecated.base-decorators-path';
 
 export default {
-  // Aliases
-  'deprecated-4-generate-uid': deprecatedGenerateUid,
-  'deprecated-4-deep-compare': deprecatedDeepCompare,
-  'deprecated-4-event-utils': deprecatedEventUtils,
-  'deprecated-4-traversing-query': deprecatedTraversingQuery,
-  'deprecated-4-toggleable-action-params': deprecatedToggleableActionParams,
-  'deprecated-4-media-rule-list-parse': deprecatedMediaRuleListParse,
-  // Paths
-  'deprecated-4-base-decorators-path': deprecatedBaseDecoratorsPath
+  'deprecated-4/aliases': deprecatedAliases,
+  'deprecated-4/methods': deprecatedMethods,
+  'deprecated-4/base-decorators-path': deprecatedPaths
 };
