--- eslint-plugin/src/rules/4/deprecated.deep-compare.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `deepCompare` alias for {@link isEqual}
- */
-export default buildRule({
-  alias: 'isEqual',
-  deprecation: 'deepCompare'
-});
