--- eslint-plugin/src/rules/4/deprecated.event-utils.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `EventUtils` alias for {@link ESLEventUtils}
- */
-export default buildRule({
-  alias: 'ESLEventUtils',
-  deprecation: 'EventUtils'
-});
