--- eslint-plugin/src/rules/4/deprecated.generate-uid.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `generateUId` alias for {@link randUID}
- */
-export default buildRule({
-  alias: 'randUID',
-  deprecation: 'generateUId'
-});
