--- eslint-plugin/src/rules/4/deprecated.toggleable-action-params.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `ToggleableActionParams` alias for {@link ESLToggleableActionParams}
- */
-export default buildRule({
-  alias: 'ESLToggleableActionParams',
-  deprecation: 'ToggleableActionParams'
-});
