--- eslint-plugin/src/rules/4/deprecated.traversing-query.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `TraversingQuery` alias for {@link ESLTraversingQuery}
- */
-export default buildRule({
-  alias: 'ESLTraversingQuery',
-  deprecation: 'TraversingQuery'
-});
