+++ eslint-plugin/src/rules/5/all.rules.ts
@@ -1,11 +1,5 @@
-import deprecatedAlertActionParams from './deprecated.alert-action-params';
-import deprecatedPanelActionParams from './deprecated.panel-action-params';
-import deprecatedPopupActionParams from './deprecated.popup-action-params';
-import deprecatedTooltipActionParams from './deprecated.tooltip-action-params';
+import deprecatedAliases from './deprecated.aliases';
 
 export default {
-  'deprecated-5-alert-action-params': deprecatedAlertActionParams,
-  'deprecated-5-panel-action-params': deprecatedPanelActionParams,
-  'deprecated-5-popup-action-params': deprecatedPopupActionParams,
-  'deprecated-5-tooltip-action-params': deprecatedTooltipActionParams
+  'deprecated-5/aliases': deprecatedAliases,
 };
