--- eslint-plugin/src/rules/5/deprecated.alert-action-params.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `AlertActionParams` alias for {@link ESLAlertActionParams}
- */
-export default buildRule({
-  alias: 'ESLAlertActionParams',
-  deprecation: 'AlertActionParams'
-});
