+++ eslint-plugin/src/rules/5/deprecated.aliases.ts
@@ -1,0 +1,20 @@
+import {buildRule} from '../../core/deprecated-alias';
+
+/**
+ * Rule for deprecated aliases
+ */
+export default buildRule([
+  {
+    alias: 'ESLAlertActionParams',
+    deprecation: 'AlertActionParams'
+  }, {
+    alias: 'ESLPanelActionParams',
+    deprecation: 'PanelActionParams'
+  }, {
+    alias: 'ESLPopupActionParams',
+    deprecation: 'PopupActionParams'
+  }, {
+    alias: 'ESLTooltipActionParams',
+    deprecation: 'TooltipActionParams'
+  }
+]);
