--- eslint-plugin/src/rules/5/deprecated.panel-action-params.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `PanelActionParams` alias for {@link ESLPanelActionParams}
- */
-export default buildRule({
-  alias: 'ESLPanelActionParams',
-  deprecation: 'PanelActionParams'
-});
