--- eslint-plugin/src/rules/5/deprecated.popup-action-params.ts
@@ -1,9 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `PopupActionParams` alias for {@link ESLPopupActionParams}
- */
-export default buildRule({
-  alias: 'ESLPopupActionParams',
-  deprecation: 'PopupActionParams'
-});
