--- eslint-plugin/src/rules/5/deprecated.tooltip-action-params.ts
@@ -1,10 +1,0 @@
-import {buildRule} from '../../core/deprecated-alias';
-
-/**
- * Rule for deprecated `TooltipActionParams` alias for {@link ESLTooltipActionParams}
- */
-export default buildRule({
-  alias: 'ESLTooltipActionParams',
-  deprecation: 'TooltipActionParams'
-});
-
